requirements.txt
googlenet_pytorch==0.3.0
googletrans==3.0.0
matplotlib==3.5.1
numpy==1.24.3
opencv_python==4.9.0.80
pandas==1.4.1
Pillow==9.0.1
Pillow==10.3.0
pip==23.3.1
psutil==5.9.5
scikit_learn==1.2.2
torch==2.2.2
torchvision==0.17.2
transformers==4.41.2
